Moved raw handling to rawhandle.py
All commands like getCommand and getChannel will now be in the file
rawhandle.py for this is where the raw text from the irc is handled to
return usable strings.
